refactor(test): move custom zcash tests
Most of them don't seem to be custom apart from the `network`
parameter.

The tests can probably be used as combined tests with a separate
change.
