+++ test/block.js
@@ -2,7 +2,6 @@
 
 var assert = require('assert')
 var Block = require('../src/block')
-var NETWORKS = require('../src/networks')
 
 var fixtures = require('./fixtures').combine('block', ['zcash'])
 
@@ -48,35 +47,6 @@ describe('Block', function () {
         }, new RegExp(f.exception))
       })
     })
-
-    fixtures.zcash.valid.forEach(function (f) {
-      it('Zcash imports ' + f.description, function () {
-        var block = Block.fromHex(f.hex, NETWORKS.zcashTest)
-
-        assert.strictEqual(block.version, f.version)
-        assert.strictEqual(block.prevHash.reverse().toString('hex'), f.prevHash)
-        assert.strictEqual(block.merkleRoot.reverse().toString('hex'), f.merkleRoot)
-        assert.strictEqual(block.timestamp, f.timestamp)
-        assert.strictEqual(block.bits, parseInt(f.bits, 16))
-        assert.strictEqual(block.nonce.reverse().toString('hex'), f.nonce)
-        assert.strictEqual(block.finalSaplingRoot.reverse().toString('hex'), f.finalSaplingRoot)
-        assert.strictEqual(block.solutionSize, f.solutionSize)
-        assert.strictEqual(block.solution.toString('hex'), f.solution)
-
-        assert.strictEqual(block.transactions.length, f.transactions.length)
-        for (var t = 0; t < f.transactions.length; t++) {
-          assert.strictEqual(block.transactions[t].ins.length, f.transactions[t].ins.length)
-          assert.strictEqual(block.transactions[t].outs.length, f.transactions[t].outs.length)
-
-          for (var i = 0; i < f.transactions[t].ins.length; i++) {
-            assert.strictEqual(block.transactions[t].ins[i].hash.toString('hex'), f.transactions[t].ins[i].hash)
-          }
-          for (var o = 0; o < f.transactions[t].outs.length; o++) {
-            assert.strictEqual(block.transactions[t].outs[o].value, f.transactions[t].outs[o].value)
-          }
-        }
-      })
-    })
   })
 
   describe('toBuffer/toHex', function () {
@@ -89,19 +59,6 @@ describe('Block', function () {
 
       it('exports ' + f.description, function () {
         assert.strictEqual(block.toHex(true), f.hex.slice(0, 160))
-        assert.strictEqual(block.toHex(), f.hex)
-      })
-    })
-
-    fixtures.zcash.valid.forEach(function (f) {
-      var block
-
-      beforeEach(function () {
-        block = Block.fromHex(f.hex, NETWORKS.zcashTest)
-      })
-
-      it('Zcash exports ' + f.description, function () {
-        assert.strictEqual(block.toHex(true), f.hex.slice(0, Block.ZCASH_HEADER_BYTE_SIZE * 2))
         assert.strictEqual(block.toHex(), f.hex)
       })
     })
@@ -120,19 +77,6 @@ describe('Block', function () {
         assert.strictEqual(block.getId(), f.id)
       })
     })
-
-    fixtures.zcash.valid.forEach(function (f) {
-      var block
-
-      beforeEach(function () {
-        block = Block.fromHex(f.hex, NETWORKS.zcashTest)
-      })
-
-      it('Zcash returns ' + f.id + ' for ' + f.description, function () {
-        assert.strictEqual(block.getHash().reverse().toString('hex'), f.hash)
-        assert.strictEqual(block.getId(), f.id)
-      })
-    })
   })
 
   describe('getUTCDate', function () {
@@ -144,20 +88,6 @@ describe('Block', function () {
       })
 
       it('returns UTC date of ' + f.id, function () {
-        var utcDate = block.getUTCDate().getTime()
-
-        assert.strictEqual(utcDate, f.timestamp * 1e3)
-      })
-    })
-
-    fixtures.zcash.valid.forEach(function (f) {
-      var block
-
-      beforeEach(function () {
-        block = Block.fromHex(f.hex, NETWORKS.zcashTest)
-      })
-
-      it('Zcash returns UTC date of ' + f.id, function () {
         var utcDate = block.getUTCDate().getTime()
 
         assert.strictEqual(utcDate, f.timestamp * 1e3)
@@ -185,20 +115,6 @@ describe('Block', function () {
         assert.strictEqual(Block.calculateMerkleRoot(block.transactions).toString('hex'), f.merkleRoot)
       })
     })
-
-    fixtures.zcash.valid.forEach(function (f) {
-      if (f.hex.length === 160) return
-
-      var block
-
-      beforeEach(function () {
-        block = Block.fromHex(f.hex, NETWORKS.zcashTest)
-      })
-
-      it('Zcash returns ' + f.merkleRoot + ' for ' + f.id, function () {
-        assert.strictEqual(Block.calculateMerkleRoot(block.transactions).reverse().toString('hex'), f.merkleRoot)
-      })
-    })
   })
 
   describe('checkMerkleRoot', function () {
@@ -215,20 +131,6 @@ describe('Block', function () {
         assert.strictEqual(block.checkMerkleRoot(), true)
       })
     })
-
-    fixtures.zcash.valid.forEach(function (f) {
-      if (f.hex.length === 160) return
-
-      var block
-
-      beforeEach(function () {
-        block = Block.fromHex(f.hex, NETWORKS.zcashTest)
-      })
-
-      it('Zcash returns ' + f.valid + ' for ' + f.id, function () {
-        assert.strictEqual(block.checkMerkleRoot(), true)
-      })
-    })
   })
 
   describe('checkProofOfWork', function () {
@@ -240,18 +142,6 @@ describe('Block', function () {
       })
 
       it('returns ' + f.valid + ' for ' + f.id, function () {
-        assert.strictEqual(block.checkProofOfWork(), f.valid)
-      })
-    })
-
-    fixtures.zcash.valid.forEach(function (f) {
-      var block
-
-      beforeEach(function () {
-        block = Block.fromHex(f.hex, NETWORKS.zcashTest)
-      })
-
-      it('Zcash returns ' + f.valid + ' for ' + f.id, function () {
         assert.strictEqual(block.checkProofOfWork(), f.valid)
       })
     })
